Minor rework to how time engine and area engine work to make logic more consistent and easier to work with. Idle hours and used hours are now separate, so they'll work more effectively. It is now only possible to set an area as delayed via a time related action, which means bad logic there is removed. Helpers for dealing with unique time based logic added.
Frontend recieved addtional areas, most only partially implemented. later commits will finish these.

+++ CoC.Backend/Delegates/Delegates.cs
@@ -29,5 +29,11 @@ namespace CoC.Backend
 	//every function that uses a button will follow this format. note that with lambdas they can take virtually anything.
 	public delegate void PlayerFunction(Player player);
 
-	public delegate void ResumeTimeCallback(byte hoursRemaining, AreaBase currentLocation);
+	/// <summary>
+	/// Used to 
+	/// </summary>
+	/// <param name="hoursRemaining"></param>
+	/// <param name="currentLocation"></param>
+	/// <param name="isIdling"></param>
+	public delegate void ResumeTimeCallback(ushort hoursRemaining, AreaBase currentLocation);
 }
